ai_project_manager/requirements.txt
streamlit
watchdog
google-generativeai